+++ Pandora/Paradigm/Junction.hs
@@ -1,3 +1,0 @@
-module Pandora.Paradigm.Junction (module Exports) where
-
-import Pandora.Paradigm.Junction.Schemes as Exports