+++ Pandora/Paradigm/Junction/Schemes.hs
@@ -1,7 +1,0 @@
-module Pandora.Paradigm.Junction.Schemes (module Exports) where
-
-import Pandora.Paradigm.Junction.Schemes.UTU as Exports
-import Pandora.Paradigm.Junction.Schemes.UT as Exports
-import Pandora.Paradigm.Junction.Schemes.TUVW as Exports
-import Pandora.Paradigm.Junction.Schemes.TUV as Exports
-import Pandora.Paradigm.Junction.Schemes.TU as Exports