+++ Pandora/Pattern/Junction.hs
@@ -1,4 +1,9 @@
 module Pandora.Pattern.Junction (module Exports) where
 
+import Pandora.Pattern.Junction.Schemes.UTU as Exports
+import Pandora.Pattern.Junction.Schemes.UT as Exports
+import Pandora.Pattern.Junction.Schemes.TUVW as Exports
+import Pandora.Pattern.Junction.Schemes.TUV as Exports
+import Pandora.Pattern.Junction.Schemes.TU as Exports
 import Pandora.Pattern.Junction.Transformer as Exports
 import Pandora.Pattern.Junction.Composition as Exports