+++ CHANGELOG.md
@@ -98,5 +98,4 @@
 * Rename `composition` to `unwrap` and `Outline` to `Primary` in `Composition` typeclass
 * Rename `equip` to `wrap` and `Layout` to `Schema` in `Transformer` typeclass
 * Make `Composition` a superclass for `Transformer` typeclass
-* Move `Composition` and `Transformer` typeclasses to `Pattern` submodule
-* Move `Kan` module to `Basis` submodule
+* Move all `Junction` modules to `Pattern` submodule except `Kan`